Move duplicate review before preprocess
